requirements.txt
azure-ai-textanalytic
